requirements.txt
flask
joblib
pandas
numpy
mysql-connector-python
python-dotenv